Fixed PHP-623: Make MD5 functions static, and only visible in the utils.c file.
We did not properly name-space protect the MD5_* functions. In case they were
already defined the linker might pick up the wrong ones. That is likely to
happen in case libcrypto is also available in the process' address space. By
making them static and removing all references in the .h file, their use in the
driver is contained.