Improve existency and free checks.
We should reset the parameters to NULL at every iteration of the loop as free()
does not set the ptr to NULL. This might technically not be needed as the hash
break up function sets things to NULL, but better safe than sorry.  The ->db
argument should also be checked for as, we do just strcmp() it later.