Fixed bug #644: mongo.ping_interval and mongo.is_master_interval is unused.
The problem was not that they were not used, but only that ini_get() / php -i
didn't show the set value. And this was because we didn't return anything from
the callback functions. We need to return SUCCESS in order for PHP to update
it's internal cache of these values.