Fixed MongoCursor->slaveOkay(false) and added support for MongoCursor::$slaveOkay.
In order to support slaveOkay(false), I've moved the parameter parsing into the
slaveOkay method and substituted the -1 (which means that parameters would be
parsed in set_cursor_flag too) for the value. If I wouldn't have done that, the
parameters would have to be parsed twice.

In mongo_cursor__do_query I moved setting the cursor opts to below the
new check for the static property value of MongoCursor::$slaveOkay, so that
its value is also taken into account.

I've also added a bunch of extra tests!