Fixed PHP-355: Driver does not honour connection 'timeout'
Previous fix to PHP-356 fixed the timeout for creating the actual
connection, while this fix deals with timeout for the ping/pong game we
need to play with mongod when bootstrapping the connection.

Note: These are two (three) different options!
connectTimeoutMS is exclusively used for connecting to the server
socketTimeoutMS  is used for data transmissions
and the 3rd one
wTimeoutMS       is _server side_ replication timeout

The socketTimeoutMS can be adjusted per-query, while connectTimeoutMS
can only be set in the ctor.
The connectTimeoutMS is is also used when attempting to reestablish
connection to the server after f.e. query timeout (socketTimeoutMS).

To make things consistent we should probably consider allowing
"socketTimeoutMS" as a query options, today it is simply called
"timeout" which is a littlebit ambiguous