Fix this annoying test
Man. We really aren't testing replication, or how fast these chunks get
replicated to the server we hit. We cannot use tagged WriteConcern and
ReadPreferenceTags as this test is meant to test slaveOkay()
exclusively, which tags would mess up. Rather then verifying the output
of the queries/commands, we just need to know that we hit the
secondaries