Fixed PHP-801: Deprecate boolean options to MongoCollection::insert().
This removes creating artificial structures in insert(), which should make it
use less memory and speed it up. However, it pushes the checking on whether
options were passed to the places where they were used. In collection.c, there
is already another places where we do "if (options &&". log_stream.c is now
required to create an artificial structure when passing things to the logging
handler. In case no options were passed, that used to show an array(), but I
think it makes more sense to use NULL as no options (as opposed to an empty
array) were passed. This caused a test (tests/mongos/stream-logger-001.phpt) to
fail as it expected array() instead of NULL.