Fixed PHP-809: Deprecate use of "safe" all over collection.c.
We only have to check and warn for the old "safe" in is_gle_op, but not again
in append_getlasterror as then it would always just show up twice.

I've changed all the "safe" to "w" in the tests, even in places where due to
MongoClient this isn't necessary and they instead can be removed. However, I
found that out of the scope of this PR and it's probably better to be implicit
about them anyway.