Fixed PHP-796: Modifying MongoDate internal properties evilness
Also fixes PHP-797: Deprecate public properties

We only raise the deprecation warning when the user is trying to set
these properties. There seems to be a lot of code in the wild that reads
the properties and as that doesn't do us any harm we shouldn't deprecate
that, or disallow it.
However, setting these properties from userland creates all sorts of
validation problems for us, and dozens of dozens of different segfaults.

Note that we still set these properties internally, which does not raise
the warning.