Fixed PHP-423: MongoDate support for 64-bit timestamps on 32-bit systems.
I can't create a test case as there is no way to create a "larger" timestamp on
32 bit platforms in the first place. After this change, MongoDate will also
show that MongoDB really only supports milliseconds, and not microseconds as
MongoDate implies. This means that a var_dump just after creation will truncate
to three decimals. This is a tiny BC break, but IMO it is more correct to do it
like this, as it reflects how it really works.