Fixed data types for INI entries.
Right, so this cost me an hour or two. The "inc" value (next commit) would
always reset to 0 when calling register_ini_entries, only because the struct's
item before hand would write 8 bytes instead of the allocated 4. What a pain
:-) It also changes the return values in the stream-logger-001 test case. The
values are stable, but were just reset to 0 and hence there are now new
numbers.