Fixed PHP-874: New Replica Set connections hang when secondaries are creating indexes during initialSync
When checking the node type, we should drop unknown mbers right away!

If we don't drop the connection here we will continue and try to
authenticate - which will attempt to take a lock on the server..
which will go so well during initial sync and index creations for
example.

This was "only" a problem when using authentication, as even if we
had the connection around normally, we don't have read preferences
that would use it so no harm no faul, just a silly thing to keep around