Fixes PHP-888: Restore 1.4.x DBRef creation behavior
This restores legacy behavior for creating DBRefs, which was changed in 926d08d6965e88a07e2c9c0cd029484142a79c4e. Tests have been added for the following behavior, which may be changed in the future.

MongoDBRef::create() should not process or validate its parameters; they should be used as-is in the resulting DBRef array. The collection and DB helper methods should attempt to resolve an array/object $id parameter to its _id field. Null should be returned if, and only if, the _id field of an array is not found. Objects that cannot be resolved to an _id property should be returned as-is. No errors should be emitted by any of these methods.

Lastly, this commit also moves the MongoDBRef::create() test (originally created in #509) to the no-servers directory.