PHP-892: Add rudimentary master/slave support
It is detected as a "MULTIPLE", since there is no concept of a set name.
The master is detected as a standalone node as we can't differnciate it
from a standalone connection (without adding a new "master/slave" option
to the $options array), but the slave will be marked as a secondary node
(without replicaset).

When getting a writable connection we will filter out the slave, but for
read connections we simply select the "nearest" node.