PHP-923: Bail out on unsupported Wire Versions
When we hit a server running unsupported wire version we don't abort
right away. Instead, we continue our ReplicaSet discovery as usual
and abort before returning a connection.
This is done incase there are multiple errors that we need to
aggregate for the user, so he doesn't have to fix one server,
retry, fix another server and so on.