Write the data to all 4 nodes
This test is testing wrong readPreferenceTags (e.g. not existing)
There is no need to be pretend the write may not have reached all nodes
and get lucky that it is on the dc:ny node.
In fact, this sometimes fails on travis as the write gets replicated
to dc:sf, dc:sf (and master dc:ny). When we then do
secondary reads & only dc:ny...
The secondary node dc:ny didn't get the write yet......

This fixes a random failure on Travis and Jenkins