Match any node type
These tests are simply testing if we spit out log entries.
It is not testing the inner workings of the ranking logic.
It is irrelevant if a server or two is lagging here, there is no
to enforce the primary to be the first server in the list here.
During ctor, all we care about is having a readable node, so the sort
is totally based on "lag", not type in this particular case.

This fixes these tests failing "randomly" on both Travis and Jenkins