Added the connection hash that was used while running a command to its output.
We are adding this as a subitem of the $php key, as we can not rely on f.e.
"hash" to be added by the server at some point. Hence we namespace with $php.
It also adds this for *every* command to be consistent. Of course, right now,
it's going to be useful only for MongoCommandCursor::fromDocument, which is
coming in the follow up PR.