Make sure we can talk to an arbiter with the rest of the nodes down.
It's near impossible to make a reliable automated test for this. I've tested
the following manually:

- make servers
- put all the secondary nodes in recovery mode with replSetMaintenance: 1
- step down the primary
- put the remaining (now secondary node) in recovery mode

Then requesting a standalone connection to the arbiter, or a replicaset
connection to the arbiter work just fine. Of course, upon writing I get a "not
master" error (which apparently changed code in 2.6 from 10058 to 100).