Differenciate between last_ismaster and a ReplicaSet verification check.
We don't need to do two isMaster() checks for mongos/standalone connections
But we do need it for a ReplicaSet. The second call does the ReplicaSet integrity check,
such as match the ReplicaSet name and who "we" are, along with extracting out tags.

We need a way to be able to run this check even if we have called isMaster() as we don't
cache the actual responses.