PHP-1160: Support MongoCollection::count() $options array
Scalar limit/skip option arguments are still supported, but $options array will allow "hint" option to be specified.

Require an array or object for MongoCollection::count() $query arg. Previously, an invalid type for this argument was ignored and passed through to the server (where it was likely ignored). The lack of validation may have been an oversight, or done purposely so that null could be specified as a convenience if the user only wanted to provide limit/skip scalar arguments.