PHP-953: Do not check custom or generated index name lengths
The 127-byte limit on index names is specific to mmapv1 storage. For both commands and legacy inserts (to system.indexes), we can rely on the server to validate the length accordingly.

Added tests for legacy inserts to ensure an exception is still thrown (albeit with a different message directly from the server).