PHP-1365: Validate cursor command option, but don't enforce batchSize
The driver should not enforce the cursor.batchSize command option, as the server will already use its own default. We should validate that the cursor option is an array/object, and empty arrays should be converted to an object to ensure proper BSON serialization (the server requires an object type).

Only aggregateCursor() requires that we inject the cursor option if it is missing. Collection and index enumeration commands do not require the option at all, as they will always return cursors in 2.8.0-RC4+ and return inline results in earlier versions. RC3 can return an emulated cursor response with only a first batch, but it is practically equivalent to the result when the cursor option is not specified at all.