PHP-1367: Restore read pref after forcing primary for variadic aggregate()
Previously, the variadic form of aggregate() would force a primary read pref if the pipeline contained $out, but never restore it. This extracts the check for $out to its own function and moves it to php_mongodb_aggregate(), which is called by both the variadic and non-variadic forms of aggregate().

Also, we need only check for $out in the last pipeline element, since the server does not allow it elsewhere.