PHP-1427 and PHP-1433: distinct() query param handling
An empty query parameter should be omitted from the command document (consistent with other helpers). This will avoid the case where an empty array is serialized as a BSON array and the server complains that a BSON object was not provided.

Additionally, relax ZPP parsing to allow arrays or objects for the query (also consistent with other helpers).