PHP-1426: Reading BSON key should not take us to buffer's end
When reading the document key, we should ensure that we're still ahead of the buffer's end. Not every BSON type has a payload and will call CHECK_BUFFER_LEN() (e.g. null, MinKey), and we shouldn't risk reading the next field's type byte if it would be out of range.