PHP-1460: Advancing cursor beyond its limit should kill it
Previously, a cursor advanced beyond limit would simply be marked as dead (i.e. dead boolean set to true, cursor ID and connection cleared) without killing the cursor on the server. This was inconsistent with behavior in hasNext(), which was refactored for 1.6.0.

In addition to killing the cursor, we should still clear its ID (this ensures MongoCursorInterface::dead() still functions properly).

Elsewhere, php_mongo_cursor_mark_dead() is only used as an error callback when we cannot find a suitable connection for issuing a query or initializing a command cursor.